Before
```
goos: darwin
goarch: amd64
pkg: github.com/damnever/skiplist
BenchmarkPut-4           2000000               880 ns/op             489 B/op          4 allocs/op
BenchmarkGet-4           5000000               405 ns/op               8 B/op          1 allocs/op
PASS
ok      github.com/damnever/skiplist    11.253s
```

After
```
goos: darwin
goarch: amd64
pkg: github.com/damnever/skiplist
BenchmarkPut-4           3000000               533 ns/op              73 B/op          3 allocs/op
BenchmarkGet-4           5000000               395 ns/op               8 B/op          1 allocs/op
PASS
ok      github.com/damnever/skiplist    8.122s
```